#105 Rename machine_learning to forecasting folder
