+++ tests/sdk/python/rtdip_sdk/pipelines/forecasting/spark/test_k_nearest_neighbors.py
@@ -22,7 +22,7 @@ from pyspark.sql.types import (
     FloatType,
 )
 from datetime import datetime
-from src.sdk.python.rtdip_sdk.pipelines.machine_learning.spark.k_nearest_neighbors import (
+from src.sdk.python.rtdip_sdk.pipelines.forecasting.spark.k_nearest_neighbors import (
     KNearestNeighbors,
 )
 from pyspark.ml.feature import VectorAssembler, StandardScaler, StringIndexer
